+++ src/main/java/gg/projecteden/nexus/features/events/y2021/bearfair21/quests/resources/WoodCutting.java
@@ -5,7 +5,7 @@ import gg.projecteden.nexus.Nexus;
 import gg.projecteden.nexus.features.events.y2021.bearfair21.BearFair21;
 import gg.projecteden.nexus.framework.exceptions.postconfigured.InvalidInputException;
 import gg.projecteden.nexus.models.cooldown.CooldownService;
-import gg.projecteden.nexus.models.job.jobs.BearFair21TreeRegenJob;
+import gg.projecteden.nexus.models.scheduledjobs.jobs.BearFair21TreeRegenJob;
 import gg.projecteden.nexus.utils.ItemBuilder;
 import gg.projecteden.nexus.utils.ItemUtils;
 import gg.projecteden.nexus.utils.PlayerUtils;
